+++ cluster/lib/src/cluster.dart
@@ -27,7 +27,7 @@ typedef AfterClusterInit = FutureOr<void> Function(
 /// Called to preprare the actor system on every worker of a node. Use this
 /// callback to register actor factories.
 typedef PrepareNodeSystem = FutureOr<void> Function(
-    void Function(Pattern pattern, ActorFactory factory) registerFactory);
+    void Function(PathMatcher pathMatcher, ActorFactory factory) addActorFactory);
 
 enum NodeState {
   created,