+++ system/lib/actor_system.dart
@@ -1,6 +1,7 @@
 export 'src/system/actor.dart' show Actor, ActorFactory, ExternalActorCreate, ExternalActorLookup;
 export 'src/system/base.dart' show actorScheme, localSystem, actorPath, localActorPath;
-export 'src/system/context.dart' show BaseContext, ActorContext, ActorSystem, MissingHostHandling;
-export 'src/system/exceptions.dart' show SkipMessage, MessageNotDelivered, ActorStopped, MailboxFull;
+export 'src/system/context.dart'
+    show BaseContext, ActorContext, ActorSystem, MissingHostHandling, PathMatcher, patternMatcher;
+export 'src/system/exceptions.dart' show SkipMessage, MessageNotDelivered, ActorStopped, MailboxFull, NoFactoryFound;
 export 'src/system/messages.dart' show initMsg, shutdownMsg;
 export 'src/system/ref.dart' show ActorRef;