+++ system/lib/actor_system_helper.dart
@@ -1,1 +1,2 @@
 export 'src/helper/actor_builder.dart' show ActorBuilder, MessageNotHandledException;
+export 'src/helper/context_extension.dart' show ActorContextExtension;